Split source codes by classes.
